fix(#1595): separate class for custom data
