+++ src/transdialog.h
@@ -21,16 +21,45 @@
 #ifndef MM_EX_TRANSDIALOG_H_
 #define MM_EX_TRANSDIALOG_H_
 
+#include "mmcustomdata.h"
 #include "defs.h"
 
 #include "model/Model_Checking.h"
 #include "model/Model_Payee.h"
+#include "model/Model_CustomFieldData.h"
 #include "model/Model_Splittransaction.h"
 
 #include <wx/spinbutt.h>
 
+enum
+{
+    /* Transaction Dialog */
+    ID_DIALOG_TRANS_TYPE = wxID_HIGHEST + 900,
+    ID_DIALOG_TRANS_TEXTNUMBER,
+    ID_DIALOG_TRANS_BUTTONDATE,
+    ID_DIALOG_TRANS_TEXTNOTES,
+    ID_DIALOG_TRANS_TEXTAMOUNT,
+    ID_DIALOG_TRANS_TOTEXTAMOUNT,
+    ID_DIALOG_TRANS_STATIC_PAYEE,
+    ID_DIALOG_TRANS_BUTTONPAYEE,
+    ID_DIALOG_TRANS_BUTTONTO,
+    ID_DIALOG_TRANS_STATUS,
+    ID_DIALOG_TRANS_ADVANCED_CHECKBOX,
+    ID_DIALOG_TRANS_ADVANCED_FROM,
+    ID_DIALOG_TRANS_ADVANCED_TO,
+    ID_DIALOG_TRANS_SPLITCHECKBOX,
+    ID_DIALOG_TRANS_BUTTONTRANSNUM,
+    ID_DIALOG_TRANS_PAYEECOMBO,
+    ID_DIALOG_TRANS_BUTTON_FREQENTNOTES,
+    ID_DIALOG_TRANS_DATE_SPINNER,
+    ID_DIALOG_TRANS_CUSTOMFIELDS,
+    ID_CUSTOMFIELD,
+};
+
 class wxDatePickerCtrl;
 class mmTextCtrl;
+class wxString;
+class mmCustomData;
 
 class mmTransDialog : public wxDialog
 {
@@ -61,18 +90,16 @@ public:
     );
 
     void SetDialogTitle(const wxString& title);
-    int getAccountID() { return m_trx_data.ACCOUNTID; }
-    int getToAccountID() { return m_trx_data.TOACCOUNTID; }
-    int getTransactionID() { return m_trx_data.TRANSID; }
+    int GetAccountID() { return m_trx_data.ACCOUNTID; }
+    int GetToAccountID() { return m_trx_data.TOACCOUNTID; }
+    int GetTransactionID() { return m_trx_data.TRANSID; }
 
 private:
+    mmCustomData* m_custom_fields;
     void CreateControls();
     void dataToControls();
-    bool validateData();
+    bool ValidateData();
     void SetEventHandlers();
-    bool SaveCustomValues();
-    void FillCustomFields(wxDialog* parent, wxBoxSizer* box_sizer
-        , const wxString& ref_type, int ref_id);
 
     void OnSplitChecked(wxCommandEvent& event);
     void OnOk(wxCommandEvent& event);
@@ -84,9 +111,8 @@ private:
     void OnAccountOrPayeeUpdated(wxCommandEvent& event);
     void OnDpcKillFocus(wxFocusEvent& event);
     void OnAutoTransNum(wxCommandEvent& event);
-    void OnMultiChoice(wxCommandEvent& event);
     void OnFrequentUsedNotes(wxCommandEvent& event);
-    void onNoteSelected(wxCommandEvent& event);
+    void OnNoteSelected(wxCommandEvent& event);
     void OnTransTypeChanged(wxCommandEvent& event);
     void OnSpin(wxSpinEvent&);
     void OnDateChanged(wxDateEvent& event);
@@ -145,30 +171,6 @@ private:
     bool skip_category_init_;
     bool skip_tooltips_init_;
 
-    enum
-    {
-        /* Transaction Dialog */
-        ID_DIALOG_TRANS_TYPE = wxID_HIGHEST + 900,
-        ID_DIALOG_TRANS_TEXTNUMBER,
-        ID_DIALOG_TRANS_BUTTONDATE,
-        ID_DIALOG_TRANS_TEXTNOTES,
-        ID_DIALOG_TRANS_TEXTAMOUNT,
-        ID_DIALOG_TRANS_TOTEXTAMOUNT,
-        ID_DIALOG_TRANS_STATIC_PAYEE,
-        ID_DIALOG_TRANS_BUTTONPAYEE,
-        ID_DIALOG_TRANS_BUTTONTO,
-        ID_DIALOG_TRANS_STATUS,
-        ID_DIALOG_TRANS_ADVANCED_CHECKBOX,
-        ID_DIALOG_TRANS_ADVANCED_FROM,
-        ID_DIALOG_TRANS_ADVANCED_TO,
-        ID_DIALOG_TRANS_SPLITCHECKBOX,
-        ID_DIALOG_TRANS_BUTTONTRANSNUM,
-        ID_DIALOG_TRANS_PAYEECOMBO,
-        ID_DIALOG_TRANS_BUTTON_FREQENTNOTES,
-        ID_DIALOG_TRANS_DATE_SPINNER,
-        ID_DIALOG_TRANS_CUSTOMFIELDS,
-        ID_CUSTOMFIELD,
-    };
 };
 
 #endif
